refactor: move OwnedOutput from 'stream' module to 'api/structs'
