+++ lib/repositories/owned_outputs_repository.dart
@@ -3,8 +3,8 @@ import 'dart:typed_data';
 
 import 'package:danawallet/extensions/api_amount.dart';
 import 'package:danawallet/generated/rust/api/structs/amount.dart';
+import 'package:danawallet/generated/rust/api/structs/owned_output.dart';
 import 'package:danawallet/generated/rust/lib.dart';
-import 'package:danawallet/generated/rust/stream.dart';
 import 'package:danawallet/repositories/database_helper.dart';
 import 'package:logger/logger.dart';
 import 'package:shared_preferences/shared_preferences.dart';
