+++ rust/src/api/structs/mod.rs
@@ -5,3 +5,4 @@ pub mod outpoint;
 pub mod recipient;
 pub mod recorded_transaction;
 pub mod unsigned_transaction;
+pub mod owned_output;
